+++ apps/opik-frontend/src/v2/constants/configuration.ts
@@ -1,0 +1,7 @@
+export enum CONFIGURATION_TABS {
+  FEEDBACK_DEFINITIONS = "feedback-definitions",
+  ENVIRONMENTS = "environments",
+  AI_PROVIDER = "ai-provider",
+  WORKSPACE_PREFERENCES = "workspace-preferences",
+  MEMBERS = "members",
+}
